chore: fixing tests; minor refinements

+++ tests/test_templates.py
@@ -42,6 +42,9 @@ def working_dir() -> Iterator[Path]:
         yield working_dir
 
         for src_dir in working_generated_root.iterdir():
+            if not src_dir.is_dir():
+                continue
+
             dest_dir = generated_root / src_dir.stem
             shutil.rmtree(dest_dir, ignore_errors=True)
             shutil.copytree(src_dir, dest_dir, dirs_exist_ok=True)
